Renamed app-components to pascal casing
