+++ src/app-components/Panel/Panel.module.css
