+++ src/app-components/Panel/Panel.tsx
@@ -10,8 +10,8 @@ import {
 } from '@navikt/aksel-icons';
 import cn from 'classnames';
 
-import { PANEL_VARIANT } from 'src/app-components/panel/constants';
-import classes from 'src/app-components/panel/Panel.module.css';
+import { PANEL_VARIANT } from 'src/app-components/Panel/constants';
+import classes from 'src/app-components/Panel/Panel.module.css';
 import { useIsMobile } from 'src/hooks/useDeviceWidths';
 
 export type PanelVariant = (typeof PANEL_VARIANT)[keyof typeof PANEL_VARIANT];
