+++ src/app-components/Panel/constants.ts
