+++ src/features/devtools/components/NodeInspector/NodeInspector.tsx
@@ -4,7 +4,7 @@ import React from 'react';
 import { Tabs } from '@digdir/designsystemet-react';
 import { Close } from '@navikt/ds-icons';
 
-import { Button } from 'src/app-components/button/Button';
+import { Button } from 'src/app-components/Button/Button';
 import reusedClasses from 'src/features/devtools/components/LayoutInspector/LayoutInspector.module.css';
 import { NodeHierarchy } from 'src/features/devtools/components/NodeInspector/NodeHierarchy';
 import { NodeInspectorContextProvider } from 'src/features/devtools/components/NodeInspector/NodeInspectorContext';
