+++ src/features/devtools/components/OpenDevToolsButton/OpenDevToolsButton.tsx
@@ -3,7 +3,7 @@ import React from 'react';
 import { CodeIcon, XMarkOctagonFillIcon } from '@navikt/aksel-icons';
 import cn from 'classnames';
 
-import { Button } from 'src/app-components/button/Button';
+import { Button } from 'src/app-components/Button/Button';
 import classes from 'src/features/devtools/components/OpenDevToolsButton/OpenDevToolsButton.module.css';
 import { useDevToolsStore } from 'src/features/devtools/data/DevToolsStore';
 import { DevToolsTab } from 'src/features/devtools/data/types';
