+++ src/layout/FileUpload/FileUploadTable/FileTableButtons.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 
 import { PencilIcon, TrashIcon } from '@navikt/aksel-icons';
 
-import { Button } from 'src/app-components/button/Button';
+import { Button } from 'src/app-components/Button/Button';
 import { ConditionalWrapper } from 'src/components/ConditionalWrapper';
 import { DeleteWarningPopover } from 'src/features/alertOnChange/DeleteWarningPopover';
 import { useAlertOnChange } from 'src/features/alertOnChange/useAlertOnChange';
