+++ src/layout/FileUploadWithTag/EditWindowComponent.tsx
@@ -4,7 +4,7 @@ import { Combobox } from '@digdir/designsystemet-react';
 import { Grid } from '@material-ui/core';
 import deepEqual from 'fast-deep-equal';
 
-import { Button } from 'src/app-components/button/Button';
+import { Button } from 'src/app-components/Button/Button';
 import { AltinnLoader } from 'src/components/AltinnLoader';
 import { isAttachmentUploaded } from 'src/features/attachments';
 import { useAttachmentsUpdater } from 'src/features/attachments/hooks';
