+++ src/layout/PrintButton/PrintButtonComponent.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 
 import type { PropsFromGenericComponent } from '..';
 
-import { Button } from 'src/app-components/button/Button';
+import { Button } from 'src/app-components/Button/Button';
 import { Lang } from 'src/features/language/Lang';
 import { ComponentStructureWrapper } from 'src/layout/ComponentStructureWrapper';
 import { LayoutPage } from 'src/utils/layout/LayoutPage';
