+++ src/layout/SimpleTable/typeguards.ts
@@ -1,6 +1,6 @@
 import type { JSONSchema7, JSONSchema7Definition } from 'json-schema';
 
-import type { FormDataObject, FormDataValue } from 'src/app-components/table/Table';
+import type { FormDataObject, FormDataValue } from 'src/app-components/Table/Table';
 
 export function isFormDataValue(value: unknown): value is FormDataValue {
   if (typeof value === 'string' || typeof value === 'number' || typeof value === 'boolean' || value === null) {
